fix: hang on payload exceed 1k
the reason for hang on payload is because we changed epoll to mio
that only support edge trigger mode, we used to rely on level trigger
to make sure we not miss some event, but for edge trigger, we must make
sure we handle the event properly

we used to handle epoll with level trigger, the epoll event will be notified
until all data was processed. that's different in edge trigger, in edge trigger
event will be only notified once there is data. If we can not process the data
we will miss the event.

this patch fix this issue by add loop around read, and read until all data was processed.
in the case of payload exceed 1k, currently we only read 1024 bytes once there is read event.
so we missed the data that exceed 1k

Signed-off-by: xujihui1985 <xujihui1985@gmail.com>